fix: refactor

+++ 150-interview/88/solution.rs
@@ -1,8 +1,8 @@
 impl Solution {
     pub fn merge(nums1: &mut Vec<i32>, m: i32, nums2: &Vec<i32>, n: i32) {
-        let mut i = m as isize - 1;
-        let mut j = n as isize - 1;
-        let mut k = (m + n) as isize - 1;
+        let mut i = m - 1;
+        let mut j = n - 1;
+        let mut k = (m + n) - 1;
 
         while j >= 0 {
             if i >= 0 && nums1[i as usize] > nums2[j as usize] {
